refactor: 페이지네이션 컴포넌트 이름 변경
