+++ src/feature/homepage/SortByLog.tsx
@@ -1,4 +1,4 @@
-import MainPagination from '@/components/Pagination/Pagination';
+import MainPagination from '@/components/Pagination/MainPagination';
 import React from 'react';
 
 const SortByLog = () => {
@@ -8,11 +8,11 @@ const SortByLog = () => {
         <div className="font-untitled text-32 h-[23px] text-[#ABAFB5]">Sort by</div>
         <div className="font-untitled text-32 h-[23px] text-[#242528]">Log</div>
       </div>
-      <div className="w-full flex mb-12">
+      <div className="flex w-full mb-12">
         <div className="w-1/2  h-[1328px] flex flex-wrap">
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -21,7 +21,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -30,7 +30,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -39,7 +39,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -48,7 +48,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -57,7 +57,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -66,7 +66,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -75,7 +75,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -86,7 +86,7 @@ const SortByLog = () => {
         <div className="w-1/2  h-[1328px] flex flex-wrap">
           <div className="w-[664px] mr-5 mb-6">
             <div className="w-full h-[550px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -95,7 +95,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -104,7 +104,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -113,7 +113,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
@@ -122,7 +122,7 @@ const SortByLog = () => {
           </div>
           <div className="w-[314px] mr-5">
             <div className="w-full h-[218px] bg-red-100 mb-2"></div>
-            <div className="w-full font-pretendard text-16 font-bold text-black">
+            <div className="w-full font-bold text-black font-pretendard text-16">
               혼자 보내는 하루, 골목골목 숨어있는 용산 원효로 카페
             </div>
             <div className="w-full font-pretendard text-17 font-normal text-[#ABAFB5]">
